+++ py/desisurvey/forecast.py
@@ -2,12 +2,15 @@
 """
 from __future__ import print_function, division, absolute_import
 
+import os
 import collections
+import datetime
 
 import numpy as np
 
-import astropy.table
+from astropy.table import Table
 import astropy.units as u
+from astropy.time import Time
 
 import desimodel.io
 
@@ -15,6 +18,7 @@ import desisurvey.config
 import desisurvey.ephem
 import desisurvey.etc
 import desisurvey.utils
+import desisurvey.plan
 
 
 class Forecast(object):
@@ -68,7 +72,9 @@ class Forecast(object):
             self.design_hourangle = np.asarray(design_hourangle)
         # Get weather factors.
         if weather is None:
-            self.weather = desisurvey.plan.load_weather(start_date, stop_date)
+            self.weather = desisurvey.utils.get_average_dome_closed_fractions(
+                start_date, stop_date)
+            self.weather = 1-self.weather
         else:
             self.weather = np.asarray(weather)
         if self.weather.shape != (self.num_nights,):
@@ -133,9 +139,9 @@ class Forecast(object):
         print(separator * rowlen)
 
     def set_overheads(self, update_margin=True,
-                      setup={'DARK': 200, 'GRAY': 200, 'BRIGHT': 150},
-                      split={'DARK': 100, 'GRAY': 100, 'BRIGHT':  75},
-                      dead ={'DARK':  20, 'GRAY': 100, 'BRIGHT':  10}):
+                      setup={'DARK': 200, 'GRAY': 200, 'BRIGHT': 150, 'BACKUP': 150},
+                      split={'DARK': 100, 'GRAY': 100, 'BRIGHT':  75, 'BACKUP': 150},
+                      dead ={'DARK':  20, 'GRAY': 100, 'BRIGHT':  10, 'BACKUP': 10}):
         df = self.df
         df['Setup overhead / tile (s)'] = np.array([setup[p] for p in self.tiles.programs])
         df['Cosmic split overhead / tile (s)'] = np.array([split[p] for p in self.tiles.programs])
@@ -150,8 +156,8 @@ class Forecast(object):
         self.update()
 
     def set_factors(self, update_margin=True,
-                       moon    = {'DARK': 1.00, 'GRAY': 1.10, 'BRIGHT': 1.33},
-                       weather = {'DARK': 1.22, 'GRAY': 1.20, 'BRIGHT': 1.16}):
+                       moon    = {'DARK': 1.00, 'GRAY': 1.5, 'BRIGHT': 3.6, 'BACKUP': 6},
+                       weather = {'DARK': 1.22, 'GRAY': 1.20, 'BRIGHT': 1.16, 'BACKUP': 6}):
         df = self.df
         df['Moon factor'] = np.array([moon[p] for p in self.tiles.programs])
         df['Weather factor'] = np.array([weather[p] for p in self.tiles.programs])
@@ -189,3 +195,90 @@ class Forecast(object):
             self.program_progress[progidx] = np.clip(
                 progress - ntiles_observed, 0, ntiles)
             ntiles_observed += ntiles
+
+
+def forecast_plots(surveyopsdir=None, include_backup=False, cfgfile=None):
+    from matplotlib import pyplot as p
+    if surveyopsdir is None:
+        surveyopsdir = os.environ['DESI_SURVEYOPS']
+    tmain = Table.read(os.path.join(surveyopsdir, 'ops', 'tiles-main.ecsv'))
+    exps = Table.read(os.path.join(surveyopsdir, 'ops', 'exposures.ecsv'))
+    cfg = desisurvey.config.Configuration(cfgfile)
+    ephem = desisurvey.ephem.get_ephem()
+    scheduled = ephem.get_program_hours()
+    closefracs = desisurvey.utils.get_average_dome_closed_fractions(
+        cfg.first_day(), cfg.last_day())
+    programnames = ['DARK', 'GRAY', 'BRIGHT']
+    weatheradjustedhours = [
+        (1-closefracs)* sched / getattr(cfg.conditions, prog).moon_up_factor()
+        for sched, prog in zip(scheduled, programnames)]
+    weatheradjustedhours = np.sum(weatheradjustedhours, axis=0)
+    cz = desisurvey.utils.cos_zenith(tmain['DESIGNHA']*u.deg,
+                                     tmain['DEC']*u.deg)
+    am = desisurvey.utils.cos_zenith_to_airmass(cz)
+    amfac = desisurvey.etc.airmass_exposure_factor(am)
+    dustfac = desisurvey.etc.dust_exposure_factor(tmain['EBV_MED'])
+    cost = amfac*dustfac
+    costetime = cost*desisurvey.tiles.get_nominal_program_times(
+        tmain['PROGRAM'])
+    include = tmain['IN_DESI'] != 0
+    if not include_backup:
+        include &= tmain['PROGRAM'] != 'BACKUP'
+    tileid_to_rownum = {t: i for i, t in enumerate(tmain['TILEID'])}
+    lastnight = np.zeros(len(tmain), dtype='i4')
+    for i in range(len(exps)):
+        rownum = tileid_to_rownum.get(exps['TILEID'][i], -1)
+        if rownum < 0:
+            continue
+        lastnight[rownum] = max(
+            [lastnight[rownum], int(exps['NIGHT'][i])])
+    lastday = cfg.last_day()
+    lastnight[lastnight == 0] = (
+        lastday.year*10000 + lastday.month*100 + lastday.day)
+    lastmjd = Time(['-'.join([str(n)[:4], str(n)[4:6], str(n)[6:8]])
+                    for n in lastnight]).mjd
+    startdatetime = datetime.datetime.combine(
+        cfg.first_day(), datetime.time())
+    nightind = (lastmjd - Time(startdatetime).mjd).astype('i4')
+    p.plot(
+        np.cumsum(weatheradjustedhours)/np.sum(weatheradjustedhours)*100,
+        label='weather-adjusted % of hours elapsed')
+    for i in range(5):
+        p.axvline(365*(i+1), linestyle='--', color='gray')
+    countdone = tmain['DONEFRAC'].copy()
+    countdone = np.clip(countdone, 0, 1)
+    countdone[(tmain['STATUS'] == 'done') |
+              (tmain['STATUS'] == 'obsend')] = 1
+    s = np.argsort(nightind)
+    doneetime = costetime*include*countdone
+    doneetimefrac = np.cumsum(doneetime[s])
+    doneetimefrac /= np.sum(costetime*include)
+    maxind = np.max(np.flatnonzero(countdone[s] > 0))
+    p.plot(nightind[s[:maxind]], doneetimefrac[:maxind]*100,
+           label='cost-adjusted % tiles completed')
+    p.legend()
+    p.xlabel('nights since 2021-05-14')
+    p.ylabel('Percentage complete')
+    darkfrac = (
+        np.sum(doneetime*include*(tmain['PROGRAM'] == 'DARK')) /
+        np.sum(costetime*include*(tmain['PROGRAM'] == 'DARK')))
+    brightfrac = (
+        np.sum(doneetime*include*(tmain['PROGRAM'] == 'BRIGHT')) /
+        np.sum(costetime*include*(tmain['PROGRAM'] == 'BRIGHT')))
+    overallfrac = (
+        np.sum(doneetime*include)/np.sum(costetime*include))
+    lastnight = nightind[s[maxind]]
+    elapsedfrac = (np.sum(weatheradjustedhours[:lastnight]) /
+                   np.sum(weatheradjustedhours))
+    p.text(0.95, 0.05,
+           ('adjusted hours elapsed: {:5.2%}\n'
+            'adjusted tiles done: {:5.2%}\n'
+            'adjusted dark done: {:5.2%}\n'
+            'adjusted bright done: {:5.2%}\n'
+            'implied margin: {:5.2%}').format(
+                elapsedfrac, overallfrac, darkfrac, brightfrac,
+                overallfrac/elapsedfrac - 1),
+           ha='right', bbox=dict(facecolor='white', pad=10, edgecolor='none'),
+           transform=p.gca().transAxes)
+    p.xlim(0, 5*365.25)
+    p.ylim(0, 100)